Add model validations to DogWalking

+++ db/migrate/20181118221545_create_dog_walkings.rb
@@ -4,7 +4,7 @@ class CreateDogWalkings < ActiveRecord::Migration[5.2]
       t.integer :status, :default => 0
       t.integer :pets
       t.float :price
-      t.float :duration
+      t.integer :duration
       t.float :latitude
       t.float :longitude
       t.time :start
